admin-portal-backend/requirements_vercel.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg[binary]==3.2.10
fpdf2==2.7.6
fpdf==1.7.2
python-dotenv==1.0.0
